web scraping/requirements.txt
flask
flask_restful
pyngrok
Pillow
numpy
ultralytics
roboflow
wget
selenium